fix ACL permissions build.rs — Changed from tauri_build::build() to tauri_build::try_build() with InlinedPlugin that registers all 5 Kotlin USB commands. This generates the ACL permission manifests. default.json — Added usb-audio:allow-listUsbDevices, allow-requestUsbPermission, allow-getUsbDeviceInfo, allow-openUsbDevice, allow-closeUsbDevice to the permissions list. Auto sample rate — recording.rs:263-278 — When requested_max_rate == 0 (Auto), it now passes u32::MAX to negotiate_sample_rate(), which finds the highest available rate the device supports. Previously it was calling default_input_config() which always returned 44.1kHz. For a bat recording app, Auto should always mean "give me the highest rate you've got."

+++ src-tauri/src/usb_audio.rs
@@ -355,7 +355,7 @@ mod isochronous {
             let ret = unsafe {
                 libc::ioctl(
                     fd,
-                    USBDEVFS_SUBMITURB as libc::c_ulong,
+                    USBDEVFS_SUBMITURB as libc::c_int,
                     &urb.urb as *const UsbdevfsUrb,
                 )
             };
@@ -391,7 +391,7 @@ mod isochronous {
             let ret = unsafe {
                 libc::ioctl(
                     fd,
-                    USBDEVFS_REAPURB as libc::c_ulong,
+                    USBDEVFS_REAPURB as libc::c_int,
                     &mut urb_reaped as *mut *mut UsbdevfsUrb,
                 )
             };
@@ -460,7 +460,7 @@ mod isochronous {
             // Resubmit unless cancelling
             if !cancel.load(Ordering::Relaxed) {
                 let ret = unsafe {
-                    libc::ioctl(fd, USBDEVFS_SUBMITURB as libc::c_ulong, urb_reaped)
+                    libc::ioctl(fd, USBDEVFS_SUBMITURB as libc::c_int, urb_reaped)
                 };
                 if ret == 0 {
                     balls_in_air += 1;
@@ -482,7 +482,7 @@ mod isochronous {
             unsafe {
                 libc::ioctl(
                     fd,
-                    USBDEVFS_DISCARDURB as libc::c_ulong,
+                    USBDEVFS_DISCARDURB as libc::c_int,
                     &urb.urb as *const UsbdevfsUrb,
                 );
             }
@@ -492,7 +492,7 @@ mod isochronous {
             unsafe {
                 libc::ioctl(
                     fd,
-                    USBDEVFS_REAPURBNDELAY as libc::c_ulong,
+                    USBDEVFS_REAPURBNDELAY as libc::c_int,
                     &mut urb_reaped as *mut *mut UsbdevfsUrb,
                 );
             }
